backend/bigquery-to-postgres/requirements.txt
google-cloud-bigquery==3.38.0
pandas==2.3.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.44
python-dotenv==1.2.1
db_dtypes==1.4.3

pytest==7.4.3
pytest-cov==4.1.0
